Adds `Date.warp.jump()` function.
This function accepts an object parameter when the keys are units of
time such as 'year', 'mo', 'week', 'days', 'hrs', 'minutes', 'sec', 'ms'
(and several variants of each) and the values are a number. The clock
will jump forwards (or backwards in the case of negative values) the
number of specified units. For instance: `Date.warp.jump({'d': 1,
'hour': 12})` will set the clock forward a day and a half.