Fixes IE7 `Date()` bug.
IE7 doesn't know what `Window` means, so in order to properly spit out
the date as a string when `Date()` is called instead of `new Date()`,
the `this` object is checked to see if it is a `Date` object. It also
checks to make sure it's not the `Date.warped` object, since that object
calls the warped constructor when setting the clock with
`Date.warp.clock()`.